Update main.js
Quite a few changes, tested on a decently long conversation.

Worth someone else testing before pushing to main.

Sleep now has randomless, the deletion is more humanized, it now deletes from bottom of chain and keeps it's current action always in view. So the user doesn't start scrolling and throwing it off or just thinking it's not working anymore.