feat: porting extension to Firefox
Thanks to Manifest V3 (MV3) porting Chrome extensions to Firefox is simpler. Thankfully Firefox support the `chrome.*` namespace [1] so not much change is needed in that aspect. The only change to be made for the extension to work on Firefox is to add `background.js` in `"scripts"` under the `"background"` key in the `manifest.json` as Firefox doesn't support `"service_worker"` yet [2].

References:
[1] https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Build_a_cross_browser_extension#api_namespace
[2] https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/Build_a_cross_browser_extension#background_page_and_extension_service_worker